ChallengeGenerator.Challenges.ChallengeAttrsQuery (Challenge Generator v0.1.0)

The ChallengeAttrsQuery module provides a consistent struct for creating the map attributes for the aggregation query in Mongo

Types

t()

@type t() :: %ChallengeGenerator.Challenges.ChallengeAttrsQuery{
  author: BSON.Regex.t(),
  author_edabit_id: binary(),
  edabit_id: binary(),
  min_difficulty: float(),
  min_quality: float(),
  programming_language: binary(),
  tags: [binary()],
  title: BSON.Regex.t()
}

Functions

create_match_map(challenge_attrs)

@spec create_match_map(t()) :: map()

This creates a map from the main struct, discarding the nil values

Parameters

challenge_attrs: The ChallengeAttrsQuery struct to create the map from

create_query_from_map(attrs_map)

@spec create_query_from_map(map()) :: t()

This creates the main struct based on a map

Parameters

map: the map to create the struct from