ChallengeGeneratorWeb.ChallengeController (Challenge Generator v0.1.0)

This controller handles the challenge fetching and management

Functions

index(conn, params)

If a GET is requested to the root path, we redirect to the API docs

Parameters

conn: Phoenix default connection

redirect_to_docs(conn, params)

If a GET is requested to the root path, we redirect to the API docs

Parameters

conn: Phoenix default connection

show(conn, map)

If a GET is requestes with the ID in the path, we return the challenge

Parameters

conn: Phoenix default connection

id: ID of the challenge

swagger_definitions()

swagger_path_index(route)

swagger_path_show(route)