ChallengeGeneratorWeb.ChallengeView (Challenge Generator v0.1.0)

This is where the jsons are generated based on the challenge struct

Types

json_challenge()

@type json_challenge() :: %{
  _id: String.t(),
  edabit_id: String.t(),
  author: String.t(),
  author_edabit_id: String.t(),
  raw_code: String.t(),
  difficulty: float(),
  raw_instructions: String.t(),
  raw_tests: String.t(),
  programming_language: String.t(),
  quality: float(),
  summary: String.t(),
  tags: [String.t()],
  title: String.t()
}

This is the struct the result JSON will be based on

Functions

__resource__()

The resource name, as an atom, for this view

render(template, assigns \\ %{})

@spec render(binary(), %{challenges: [ChallengeGenerator.Challenges.Challenge.t()]}) ::
  %{
    data: [json_challenge()]
  }

@spec render(binary(), %{challenge: ChallengeGenerator.Challenges.Challenge.t()}) ::
  %{
    data: json_challenge()
  }

@spec render(binary(), %{challenge: ChallengeGenerator.Challenges.Challenge.t()}) ::
  json_challenge()

Generates the JSON based on the Challenge struct

template_not_found(template, assigns)

@spec template_not_found(Phoenix.Template.name(), map()) :: no_return()

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.